requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==23.2.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
certifi==2024.6.2
charset-normalizer==3.3.2
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
contourpy==1.3.0
cycler==0.12.1
debugpy @ file:///croot/debugpy_1690905042057/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
e3nn==0.5.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock==3.14.0
fonttools==4.54.0
frozenlist==1.4.1
fsspec==2024.6.0
huggingface-hub==0.23.2
hydra-core==1.3.2
idna==3.7
importlib_resources==6.4.5
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1717717528849/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1680185408135/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
joblib==1.4.2
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1654730843242/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1710257447442/work
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opt-einsum==3.3.0
opt-einsum-fx==0.1.4
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas==2.2.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.3.0
pinnstorch==0.1.2
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1715777629804/work
POT==0.9.4
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1717583537988/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyDOE==0.3.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyitcast==1.0.4
pyparsing==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.1
pyzmq @ file:///croot/pyzmq_1705605076900/work
requests==2.32.3
rich==13.9.2
rootutils==1.0.7
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
skipatom==1.2.5
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
sympy==1.12.1
threadpoolctl==3.5.0
timm==1.0.3
torch==2.2.0
torch_cluster @ file:///home/wanglin/.conda/envs/e3nn/lib/python3.9/site-packages/torch_cluster-1.6.3%2Bpt22cu121-cp39-cp39-linux_x86_64.whl#sha256=93fe2302eb3e45dfce599b7385b247b6eec92823edf9d9eaacfd23e9af455ce6
torch_geometric==2.5.3
torch_scatter @ file:///home/wanglin/.conda/envs/e3nn/lib/python3.9/site-packages/torch_scatter-2.1.2%2Bpt22cu121-cp39-cp39-linux_x86_64.whl#sha256=248cabae7dcfba6c39233104abea2055a796dcc9f902006f61f8c123165696b6
torch_sparse @ file:///home/wanglin/.conda/envs/e3nn/lib/python3.9/site-packages/torch_sparse-0.6.18%2Bpt22cu121-cp39-cp39-linux_x86_64.whl#sha256=9353e09b1f3c1e338ac2c7f0837eeedbc6b5fe813e923c51339bc5c221f6686c
torchsde==0.2.6
torchvision==0.17.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827245914/work
tqdm==4.66.4
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
trampoline==0.1.2
triton==2.2.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717287769032/work
tzdata==2024.2
urllib3==2.2.1
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
yarl==1.9.4
zipp==3.20.2
